+++ lib/src/extensions/file_extensions.dart
@@ -4,5 +4,6 @@ extension FileExtension on File {
   String get extensionType => path.split('.').last.toLowerCase();
 
   bool get hasARBExtension => extensionType == 'arb';
+  bool get hasJSONExtension => extensionType == 'json';
   bool get hasCSVExtension => extensionType == 'csv';
 }